+++ packages/blob/src/client.ts
@@ -20,7 +20,7 @@ import { createCreateMultipartUploaderMethod } from './multipart/create-uploader
 // This types omits all options that are encoded in the client token.
 export interface ClientCommonCreateBlobOptions {
   /**
-   * Whether the blob should be publicly accessible. Support for private blobs is planned.
+   * Whether the blob should be publicly accessible.
    */
   access: 'public';
   /**
@@ -384,8 +384,8 @@ export async function handleUpload({
       const signatureHeader = 'x-vercel-signature';
       const signature = (
         'credentials' in request
-          ? request.headers.get(signatureHeader) ?? ''
-          : request.headers[signatureHeader] ?? ''
+          ? (request.headers.get(signatureHeader) ?? '')
+          : (request.headers[signatureHeader] ?? '')
       ) as string;
 
       if (!signature) {
@@ -515,3 +515,5 @@ export interface GenerateClientTokenOptions extends BlobCommandOptions {
   addRandomSuffix?: boolean;
   cacheControlMaxAge?: number;
 }
+
+export { createFolder } from './create-folder';